requirements.txt
absl-py==0.12.0
aim==2.1.3
aimrecords==0.0.7
alabaster==0.7.12
altair==4.1.0
anytree==2.8.0
applaunchservices==0.2.1
astunparse==1.6.3
async-generator==1.10
axial-positional-embedding==0.2.1
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
batchgenerators==0.20.1
bkcharts==0.2
blinker==1.4
boto==2.49.0
brotlipy==0.7.0
certifi==2020.12.5
click==7.1.2
clyent==1.2.0
contextlib2==0.6.0.post1
contrastive-learner==0.1.0
cycler==0.10.0
cytoolz==0.11.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docker==4.4.0
et-xmlfile==1.0.1
fastai==1.0.61
fire==0.3.1
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
glob2==0.7
gmpy2==2.1.0b1
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gym==0.17.3
h5py==2.10.0
HeapDict==1.0.1
imagesize==1.2.0
intervaltree==3.0.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jupyterlab==2.2.9
Keras-Preprocessing==1.1.2
kornia==0.4.1
linear-attention-transformer==0.15.0
linecache2==1.0.0
linformer==0.2.0
llvmlite==0.35.0
local-attention==1.1.1
locket==0.2.0
Markdown==3.3.4
matplotlib==3.3.3
mccabe==0.6.1
medicaltorch==0.2
mpmath==1.1.0
multipledispatch==0.6.0
nltk==3.4.4
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.5
pandocfilters==1.4.2
parso==0.7.0
partd==1.1.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
Pillow==6.2.2
plac==0.9.6
ply==3.11
product-key-memory==0.1.10
protobuf==3.14.0
ptyprocess==0.6.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0.6
pyflakes==2.2.0
pyglet==1.5.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrser==0.2.0
pytest==6.1.2
python-dateutil==2.8.1
pytorch-fid==0.1.1
PyYAML==5.3.1
pyzmq==20.0.0
QtPy==1.9.0
requests-oauthlib==1.3.0
retry==0.9.2
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
snowballstemmer==2.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stylegan2-pytorch==1.5.1
tables==3.6.1
tblib==1.6.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
testpath==0.4.4
tifffile==2019.7.26.2
torch==1.7.0
torchvision==0.8.1
traceback2==1.4.0
unicodecsv==0.14.1
unittest2==1.1.0
vector-quantize-pytorch==0.1.0
virtualenv==20.1.0
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
